Handle syntax coloring of variable parameters
