+++ newIDE/app/src/EventsSheet/ParameterFields/VariableField.js
@@ -14,6 +14,7 @@ import {
   icon,
   nameAndIconContainer,
   instructionWarningParameter,
+  instructionParameter,
 } from '../EventsTree/ClassNames';
 import SemiControlledAutoComplete, {
   type SemiControlledAutoCompleteInterface,
@@ -21,7 +22,12 @@ import SemiControlledAutoComplete, {
 } from '../../UI/SemiControlledAutoComplete';
 import { TextFieldWithButtonLayout } from '../../UI/Layout';
 import { type ParameterInlineRendererProps } from './ParameterInlineRenderer.flow';
-import { highlightSearchText } from '../../Utils/HighlightSearchText';
+import {
+  renderStylizedText,
+  mergeStylizedText,
+  getHighlightSearchTextParts,
+  applySyntaxColoring,
+} from '../../Utils/HighlightSearchText';
 import ShareExternal from '../../UI/CustomSvgIcons/ShareExternal';
 import SelectField from '../../UI/SelectField';
 import SelectOption from '../../UI/SelectOption';
@@ -642,6 +648,7 @@ export default (React.forwardRef<Props, VariableFieldInterface>(
 export const renderVariableWithIcon = (
   {
     value,
+    expression,
     parameterMetadata,
     expressionIsValid,
     hasDeprecationWarning,
@@ -651,6 +658,7 @@ export const renderVariableWithIcon = (
     projectScopedContainersAccessor,
     highlightedSearchText,
     highlightedSearchMatchCase,
+    scope,
   }: ParameterInlineRendererProps,
   tooltip: string,
   getVariableSourceFromIdentifier: (
@@ -691,11 +699,32 @@ export const renderVariableWithIcon = (
         <VariableIcon
           className={classNames({
             [icon]: true,
+            [instructionParameter]: expressionIsValid,
+            variable: true,
           })}
         />
-        {highlightSearchText(value, highlightedSearchText, {
-          matchCase: highlightedSearchMatchCase,
-        })}
+        {expressionIsValid
+          ? renderStylizedText(
+              value,
+              mergeStylizedText(
+                getHighlightSearchTextParts(value, highlightedSearchText, {
+                  matchCase: highlightedSearchMatchCase,
+                }),
+                applySyntaxColoring({
+                  text: value,
+                  expression,
+                  rootType: parameterMetadata.getValueTypeMetadata().getName(),
+                  platform: scope.project.getCurrentPlatform(),
+                  projectScopedContainers: projectScopedContainersAccessor.get(),
+                })
+              )
+            )
+          : renderStylizedText(
+              value,
+              getHighlightSearchTextParts(value, highlightedSearchText, {
+                matchCase: highlightedSearchMatchCase,
+              })
+            )}
       </IconAndNameContainer>
     </span>
   );
